fix: url 수정, transaction 목록 버그 해결

+++ chaekit-spring/src/main/java/qwerty/chaekit/dto/ebook/credit/CreditTransactionResponse.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 @Builder
 public record CreditTransactionResponse(
         String orderId,
-        Long productId,
+        int productId,
         String productName,
         CreditTransactionType type,
         int creditAmount,
